requirements.txt
python-dotenv
py-cord
logging